access self in closure
we modify the code from

	signal.subscribeNext { (e) in
	  println("number \(e)")
	}

to

	signal.subscribeNext { (e) in
	  println("number \(e) from \(self)")
	}

and then push/pop `SignalViewController`, the `SignalViewController`
still get deallocated successfully.

We verified that

1. The signal was retained by the subscriber(yes, the implicit
subscriber we create using `signal.subscribeNext`). If it is not
retained by subscriber, the signal will deallocated immediately after
alloc, because `var signal` is a weak variable, which means not
retained by `SignalViewController`
2. we guesss both `signal` and subscriber are deallocated after the
signal completed, because when we pop `SignalViewController`, and the
`signal` is nil. We will try to verify that later